removed trait objects from oks.rs and errors.rs
this new implementations should be more efficient,
and more faithful to the "zero-cost abstraction" mantra.

As a consequence, I had to move the 'process' method to a generic trait,
defined in util.rs and implemented by any Iterator.

A another consequence,
the 'static lifetime bound on the iterator can also be removed.